requirements.txt
altair==6.0.0
altex==0.2.0
asn1crypto==1.5.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
boto3==1.42.34
botocore==1.42.34
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
entrypoints==0.4
Faker==40.1.2
favicon==0.7.0
filelock==3.20.3
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
htbuilder==0.9.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lxml==6.0.2
Markdown==3.10.1
markdownlit==0.0.7
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.1
packaging==26.0
pandas==2.3.3
pillow==12.1.0
platformdirs==4.5.1
plotly==6.5.2
prometheus_client==0.24.1
protobuf==6.33.4
pyarrow==23.0.0
pycparser==3.0
pydeck==0.9.1
PyJWT==2.10.1
pymdown-extensions==10.20.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
setuptools==80.10.2
six==1.17.0
smmap==5.0.2
snowflake-connector-python==4.2.0
snowflake-snowpark-python==1.44.0
sortedcontainers==2.4.0
soupsieve==2.8.3
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.53.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-image-zoom==0.0.4
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
tenacity==9.1.2
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
validators==0.35.0
watchdog==6.0.0
wheel==0.46.3
